WIP: Got rid of EditorParagraph. Now I need to figure out the right abstraction for text selection and doc manipulation. Currently it's all shoved in EditorState.
